move deployed_contract back to test_env
