--- src/commands/test/deployed_contract.py
@@ -1,10 +1,0 @@
-from starkware.starknet.testing.contract import StarknetContract
-
-
-class DeployedContract:
-    def __init__(self, starknet_contract: StarknetContract):
-        self._starknet_contract = starknet_contract
-
-    @property
-    def contract_address(self):
-        return self._starknet_contract.contract_address
